Added to the method delete of tools the warning if it has relations with others

+++ app/controllers/herramientas.controller.js
@@ -8,7 +8,7 @@ exports.create = (req, res) => {
   if (!req.body) {
     res.status(400).json({
       status: 400,
-      message: 'Content can not be empty!'
+      message: "Content can not be empty!",
     });
     return;
   }
@@ -16,22 +16,21 @@ exports.create = (req, res) => {
   // Create a Tool
   const tool = {
     Tipo: req.body.Tipo,
-    Nombre: req.body.Nombre
+    Nombre: req.body.Nombre,
   };
-
 
   // Save Tool in the database
   Tool.create(tool)
-    .then(data => {
+    .then((data) => {
       res.status(201).json({
         status: 201,
-        message: 'Tool was created successfully!'
+        message: "Tool was created successfully!",
       });
     })
-    .catch(err => {
+    .catch((err) => {
       res.status(500).json({
         status: 500,
-        Err: err.message || 'Some error occurred while creating the Tool.'
+        Err: err.message || "Some error occurred while creating the Tool.",
       });
     });
 };
@@ -42,13 +41,13 @@ exports.findAll = (req, res) => {
   var condition = Id ? { Id: { [Op.like]: `%${Id}%` } } : null;
 
   Tool.findAll({ where: condition })
-    .then(data => {
+    .then((data) => {
       res.send(data);
     })
-    .catch(err => {
+    .catch((err) => {
       res.status(500).json({
         status: 500,
-        Err: err.message || 'Some error occurred while retrieving Tools.'
+        Err: err.message || "Some error occurred while retrieving Tools.",
       });
     });
 };
@@ -58,20 +57,20 @@ exports.findOne = (req, res) => {
   const id = req.params.id;
 
   Tool.findByPk(id)
-    .then(data => {
+    .then((data) => {
       if (data) {
         res.send(data);
       } else {
         res.status(404).json({
           status: 404,
-          message: `Cannot find Tool with id=${id}.`
+          message: `Cannot find Tool with id=${id}.`,
         });
       }
     })
-    .catch(err => {
+    .catch((err) => {
       res.status(500).json({
         status: 500,
-        message: "Error retrieving Tool with id=" + id
+        message: "Error retrieving Tool with id=" + id,
       });
     });
 };
@@ -81,54 +80,69 @@ exports.update = (req, res) => {
   const id = req.params.id;
 
   Tool.update(req.body, {
-    where: { Herramienta_id: id }
+    where: { Herramienta_id: id },
   })
-    .then(num => {
+    .then((num) => {
       if (num == 1) {
         res.status(200).json({
           status: 200,
-          message: "Tool was updated successfully."
+          message: "Tool was updated successfully.",
         });
       } else {
         res.status(400).json({
           status: 400,
-          message: `Cannot update Tool with id=${id}. Maybe Tool was not found or req.body is empty!`
+          message: `Cannot update Tool with id=${id}. Maybe Tool was not found or req.body is empty!`,
         });
       }
     })
-    .catch(err => {
+    .catch((err) => {
       console.log();
       res.status(500).json({
         status: 500,
-        message: "Error updating Tool with id=" + id
+        message: "Error updating Tool with id=" + id,
       });
     });
 };
 
+
 // Delete a Tool with the specified id in the request
-exports.delete = (req, res) => {
+exports.delete = async (req, res) => {
   const id = req.params.id;
 
-  Tool.destroy({
-    where: { Herramienta_id: id }
-  })
-    .then(num => {
-      if (num == 1) {
-        res.status(200).json({ 
-          status: 200,
-          message: "Tool was deleted successfully!"
-        });
-      } else {
-        res.status(400).json({ 
-          status: 400,
-          message: `Cannot delete Tool with id=${id}. Maybe Tool was not found!`
-        });
-      }
+  const verifyTools = await db.sequelize.query(
+    "SELECT COUNT(*) AS number FROM tb_aula_herramienta_ordenadors WHERE `Herramienta_id` = ?",
+    //"SELECT * FROM tb_aula_herramienta_ordenadors WHERE `Herramienta_id` = ?",
+    {
+      replacements: [id],
+      type: db.sequelize.QueryTypes.SELECT,
+    }
+  );
+
+  const NumberOfVerifyTools = verifyTools[0].number;
+
+  if (NumberOfVerifyTools >= 1) {
+    console.log("1");
+    Tool.destroy({
+      where: { Herramienta_id: id },
     })
-    .catch(err => {
-      res.status(500).json({ 
-        status: 500,
-        message: "Could not delete Tool with id=" + id
+      .then((num) => {
+        if (num == 1) {
+          res.status(200).json({
+            status: 200,
+            message: "Tool was deleted successfully!",
+          });
+        } else {
+          res.status(400).json({
+            status: 400,
+            message: `Cannot delete Tool with id=${id}. Maybe Tool was not found!`,
+          });
+        }
+      })
+      .catch((err) => {
+        res.status(500).json({
+          status: 500,
+          message: "Tool "+ id +" cannot be deleted because it is related to other elements.",
+        });
       });
-    });
+  }
 };
